feat: implementing validators in user

+++ src/modules/user/application/use-cases/remove-user.use-case.ts
@@ -1,6 +1,9 @@
 import { inject, injectable } from 'tsyringe';
 
+import { User } from '../../core/entities/user.entity';
 import { IUserRepository } from '../../core/repositories/user.repository';
+import { removeUserValidator } from '../validators/remove-user.validator';
+import { AppError } from '../../../../core/domain/errors/app.error';
 
 @injectable()
 export class RemoveUserUseCase {
@@ -9,8 +12,11 @@ export class RemoveUserUseCase {
     private userRepository: IUserRepository
   ) {}
 
-  async execute(id: string): Promise<void> {
-    // TODO: MAKE ONLY THE AUTHENTICATED USER CAN REMOVE ITSELF
+  async execute(id: string, user: User): Promise<void> {
+    if (!removeUserValidator(id, user)) {
+      throw new AppError('Validation error');
+    }
+
     await this.userRepository.remove(id);
   }
 }
